+++ HISTORY.md
@@ -13,6 +13,7 @@
 ### 2021-06-24
 * extensions: tweaked loading unlisted script-extensions
 * byob, threads, store: removed unused code
+* extensions: added documentation for adding external JS modules
 
 ### 2021-06-23
 * updated bignums library